IOS-7720 CR fixes
